requirements.txt
# requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
email-validator==2.1.0.post1
python-dotenv==1.0.1

requests==2.32.3
Werkzeug==3.0.1
pytz==2023.3

# Продакшен-сервер
gunicorn==21.2.0



